pyproject.toml
[tool.poetry]
name = "urlshortener"
version = "0.1.0"
description = ""
authors = ["gtx-1060 <klientkey200213@gmail.com>"]
packages = [
    { include = "app", from = "."}
]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.70.0"
SQLAlchemy = "^1.4.25"
uvicorn = "^0.15.0"
psycopg2-binary = "^2.9.1"
python-dotenv = "^0.19.0"
asyncpg = "^0.24.0"
jose = "^1.0.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
server = "app.main:start"
create_tables = "app.data.db.migrate:create_tables"
drop_tables = "app.data.db.migrate:drop_tables"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
